fix(surfaces): break resolver import cycle and preserve non-canonical persisted surfaces
Two issues the discriminated-union refactor exposed:

1. Circular import: routing the surface route resolver through the heavy
   `conversation-surfaces` module (for `restoreSurfaceStateEntry`) created
   an import cycle that broke Bun's module init order, surfacing as spurious
   "export not found" errors in `raw-query`/`env` and failing the
   surface-route test files. The surface-state types and correlated-pair
   helpers move to a standalone `conversation-surface-state` module with only
   lightweight imports; the resolver imports from there, not the heavy module.

2. Non-canonical persisted surface types: `skill_card` (memory retrospective)
   and `call_summary` (calls) are appended to history directly and were never
   in `SURFACE_TYPES`. On main the live `surfaceState` map held them via the
   casts this PR removes; removing the casts coerced them to `dynamic_page`
   and stripped their data. Both are added to the registry as opaque
   passthrough types (like `channel_setup`/`task_preferences`) so restore
   preserves them and the discriminated union stays honest. They are not in
   `SURFACE_SHAPE_DOCS`, so the `ui_show` tool still does not offer them and
   the teaching guard still rejects them.

The read-only surface content route now serves persisted `surfaceType`/`data`
verbatim (a loose read-model) rather than through the canonical-schema
restore, so it never strips keys the client renders — matching main's
behavior without reintroducing casts.

Follows up LUM-2579.
