+++ assistant/src/runtime/routes/surface-conversation-resolver.ts
@@ -14,6 +14,7 @@
  * as the existence check so `getOrCreateConversation` can't be tricked
  * into materializing a phantom conversation for any caller-supplied id.
  */
+import { coerceSurfaceDataRecord } from "../../api/surfaces.js";
 import type { Conversation } from "../../daemon/conversation.js";
 import {
   findConversation,
@@ -21,13 +22,13 @@ import {
 } from "../../daemon/conversation-registry.js";
 import { getOrCreateConversation } from "../../daemon/conversation-store.js";
 import {
-  restoreSurfaceStateEntry,
-  type SurfaceStateEntry,
-} from "../../daemon/conversation-surfaces.js";
+  parseActivationMomentTag,
+  parseStoredActions,
+  type StoredSurfaceAction,
+} from "../../daemon/conversation-surface-state.js";
 import { isRowVisibleToUntrustedActor } from "../../daemon/message-provenance.js";
-import type {} from "../../daemon/message-types/surfaces.js";
 import { rawAll, rawGet } from "../../persistence/raw-query.js";
-import {} from "../../telemetry/activation-funnel.js";
+import type { ActivationMomentParam } from "../../telemetry/activation-funnel.js";
 
 /**
  * Resolve the {@link Conversation} that owns the given surface.
@@ -79,10 +80,22 @@ export async function resolveSurfaceConversation(
 }
 
 /**
- * A `ui_surface` block extracted from persisted history, in the shape of a
- * `Conversation.surfaceState` entry.
+ * A `ui_surface` block extracted from persisted history for a read-only
+ * content response. `surfaceType` and `data` are preserved VERBATIM — this
+ * endpoint serves what was persisted so the client renders it, and the
+ * client applies the canonical per-type schema itself. It deliberately does
+ * NOT reuse the live `SurfaceStateEntry` restore (`restoreSurfaceStateEntry`),
+ * whose canonical parse strips keys the daemon does not model and coerces
+ * types it does not know (e.g. `skill_card`, `call_summary`) — which would
+ * drop renderable content on the way to the client.
  */
-export type PersistedSurfaceState = SurfaceStateEntry;
+export interface PersistedSurfaceState {
+  surfaceType: string;
+  data: Record<string, unknown>;
+  title?: string;
+  actions?: StoredSurfaceAction[];
+  activationMoment?: ActivationMomentParam;
+}
 
 /**
  * Find the `ui_surface` block for `surfaceId` in the conversation's
@@ -175,10 +188,19 @@ export function findPersistedSurfaceState(
       if (b.type !== "ui_surface" || b.surfaceId !== surfaceId) {
         continue;
       }
-      // Same field mapping and validation as
-      // `restoreSurfaceStateFromHistory` — a malformed daemon-only
-      // `activationMoment` is dropped, never served or memoized.
-      return restoreSurfaceStateEntry(b);
+      // Serve verbatim: preserve the persisted `surfaceType` string and
+      // `data` object as-is (the client parses them), applying only the same
+      // light shared parsing for `actions` and the daemon-only
+      // `activationMoment` tag that the live restore uses.
+      const activationMoment = parseActivationMomentTag(b.activationMoment);
+      return {
+        surfaceType:
+          typeof b.surfaceType === "string" ? b.surfaceType : "dynamic_page",
+        data: coerceSurfaceDataRecord(b.data),
+        title: typeof b.title === "string" ? b.title : undefined,
+        actions: parseStoredActions(b.actions),
+        ...(activationMoment ? { activationMoment } : {}),
+      };
     }
   }
   return undefined;
